Refactor authentication feature: migrate auth-related files to the new feature structure and remove obsolete code
