+++ lib/src/feature/auth/domain/entity/user_entity.dart
