+++ lib/src/feature/auth/presentation/login/cubit/login_cubit.dart
@@ -1,7 +1,7 @@
 import 'package:equatable/equatable.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:quickdrop/src/core/localization/app_localizations.dart';
-import 'package:quickdrop/src/domain/usecase/auth_usecase.dart';
+import 'package:quickdrop/src/feature/auth/domain/usecase/auth_usecase.dart';
 
 part 'login_state.dart';
