+++ lib/src/feature/auth/presentation/signup/cubit/signup_cubit.dart
@@ -1,6 +1,6 @@
 import 'package:equatable/equatable.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:quickdrop/src/domain/usecase/auth_usecase.dart';
+import 'package:quickdrop/src/feature/auth/domain/usecase/auth_usecase.dart';
 
 part 'signup_state.dart';
