+++ lib/src/feature/home/presentation/widgets/drawer/drawer.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:quickdrop/src/domain/usecase/auth_usecase.dart';
+import 'package:quickdrop/src/feature/auth/domain/usecase/auth_usecase.dart';
 import 'package:quickdrop/src/injection/injection_container.dart';
 import 'package:quickdrop/src/feature/home/presentation/widgets/drawer/widgets/custom_drawer_header/custom_drawer_header.dart';
 import 'package:quickdrop/src/feature/home/presentation/widgets/drawer/widgets/section/section_tile.dart';
