+++ lib/src/injection/injection_container.dart
@@ -2,13 +2,13 @@ import 'package:get_it/get_it.dart';
 import 'package:google_sign_in/google_sign_in.dart';
 import 'package:quickdrop/src/data/datasource/favorite_datasource.dart';
 import 'package:quickdrop/src/data/datasource/financial_information_datasource.dart';
-import 'package:quickdrop/src/data/datasource/firebase_auth_datasource.dart';
+import 'package:quickdrop/src/feature/auth/data/datasource/auth_datasource.dart';
 import 'package:quickdrop/src/feature/home/data/datasource/home_datasource.dart';
 import 'package:quickdrop/src/data/datasource/my_purchases_datasource.dart';
 import 'package:quickdrop/src/feature/productDetail/data/datasource/produt_datasource.dart';
 import 'package:quickdrop/src/data/datasource/purchase_datasource.dart';
 import 'package:quickdrop/src/data/datasource/purchase_detail_datasource.dart';
-import 'package:quickdrop/src/data/repository/auth_repository_impl.dart';
+import 'package:quickdrop/src/feature/auth/data/repository/auth_repository_impl.dart';
 import 'package:quickdrop/src/data/repository/favorite_repository_impl.dart';
 import 'package:quickdrop/src/data/repository/financial_information_repository_impl.dart';
 import 'package:quickdrop/src/feature/home/data/repository/home_data_repository_impl.dart';
@@ -16,7 +16,7 @@ import 'package:quickdrop/src/data/repository/my_purchases_repository_impl.dart'
 import 'package:quickdrop/src/feature/productDetail/data/repository/product_detail_repository_impl.dart';
 import 'package:quickdrop/src/data/repository/purchase_detail_repository_impl.dart';
 import 'package:quickdrop/src/data/repository/purchase_repository_impl.dart';
-import 'package:quickdrop/src/domain/repository/auth_repository.dart';
+import 'package:quickdrop/src/feature/auth/domain/repository/auth_repository.dart';
 import 'package:quickdrop/src/domain/repository/favorite_repository.dart';
 import 'package:quickdrop/src/domain/repository/financial_information_repository.dart';
 import 'package:quickdrop/src/feature/home/domain/repository/home_data_repository.dart';
@@ -24,7 +24,7 @@ import 'package:quickdrop/src/domain/repository/my_purchases_repository.dart';
 import 'package:quickdrop/src/feature/productDetail/domain/repository/product_detail_repository.dart';
 import 'package:quickdrop/src/domain/repository/purchase_detail_repository.dart';
 import 'package:quickdrop/src/domain/repository/purchase_repository.dart';
-import 'package:quickdrop/src/domain/usecase/auth_usecase.dart';
+import 'package:quickdrop/src/feature/auth/domain/usecase/auth_usecase.dart';
 import 'package:quickdrop/src/domain/usecase/favorite_usecase.dart';
 import 'package:quickdrop/src/domain/usecase/financial_information_usecase.dart';
 import 'package:quickdrop/src/feature/home/domain/usecase/home_data_usecase.dart';
@@ -37,11 +37,11 @@ import 'package:quickdrop/src/presentation/favorites/cubit/favorites_cubit.dart'
 import 'package:quickdrop/src/presentation/financial_information/cubit/financial_information_cubit.dart';
 import 'package:quickdrop/src/feature/home/presentation/cubit/home_cubit.dart';
 import 'package:quickdrop/src/feature/home/presentation/widgets/products/widgets/favorite/cubit/add_favorite_cubit.dart';
-import 'package:quickdrop/src/presentation/login/cubit/login_cubit.dart';
+import 'package:quickdrop/src/feature/auth/presentation/login/cubit/login_cubit.dart';
 import 'package:quickdrop/src/presentation/my_purchases/cubit/my_purchases_cubit.dart';
 import 'package:quickdrop/src/feature/productDetail/presentation/cubit/product_cubit.dart';
 import 'package:quickdrop/src/presentation/purchase_detail/cubit/purchase_detail_cubit.dart';
-import 'package:quickdrop/src/presentation/signup/cubit/signup_cubit.dart';
+import 'package:quickdrop/src/feature/auth/presentation/signup/cubit/signup_cubit.dart';
 
 final GetIt sl = GetIt.instance;
